Enhance absence management functionality; implement user-specific absence creation, retrieval, and decision-making with appropriate security checks

+++ src/main/java/tn/SmartBank/ATB_2526_SmartBank/entity/History_Sold.java
@@ -33,8 +33,8 @@ public class History_Sold {
     private LocalDate dateAction;
 
     @Column(name = "solde_before")
-    private Float soldeBefore;
+    private Double soldeBefore;
 
     @Column(name = "solde_after")
-    private Float soldeAfter;
+    private Double soldeAfter;
 }
